wip: follow show

+++ app/Http/Controllers/UserController.php
@@ -365,28 +365,54 @@ ShowRepository $showRepository)
 
         if (Request::ajax()) {
             $user = $this->userRepository->getUserByID($inputs['user_id']);
-            $show = explode(',', $inputs['shows']);
 
-            if(!isset($inputs['message'])) {
-                $message = "";
-            } else {
-                $message = $inputs['message'];
-            }
-
-            foreach($show as $s) {
-                if($user->shows->contains($s)) {
-                    $user->shows()->detach($s);
+            if(!empty($inputs['shows'])) {
+                $show = explode(',', $inputs['shows']);
+                if (!isset($inputs['message'])) {
+                    $message = "";
+                } else {
+                    $message = $inputs['message'];
                 }
-            }
 
-            $user->shows()->attach($show, ['state' => $inputs['state'], 'message' => $message]);
+                foreach ($show as $s) {
+                    if ($user->shows->contains($s)) {
+                        $user->shows()->detach($s);
+                    }
+                }
+
+                $user->shows()->attach($show, ['state' => $inputs['state'], 'message' => $message]);
+            }
 
             if($inputs['state'] == 1) {
-                $followed_shows = $this->showRepository->getShowFollowedByUser($user->id,1);
+                $followed_shows = $this->showRepository->getShowFollowedByUser($user->id);
                 $in_progress_shows = $followed_shows->where('state', '=', 1);
-                return Response::json(View::make('users.shows_cards', ['in_progress_shows' => $in_progress_shows])->render());
+                return Response::json(View::make('users.shows_cards', ['shows' => $in_progress_shows])->render());
+            } elseif ($inputs['state'] == 2) {
+                $followed_shows = $this->showRepository->getShowFollowedByUser($user->id);
+                $on_break_shows = $followed_shows->where('state', '=', 2);
+                return Response::json(View::make('users.shows_cards', ['shows' => $on_break_shows])->render());
+            } elseif ($inputs['state'] == 3) {
+                $followed_shows = $this->showRepository->getShowFollowedByUser($user->id);
+                $completed_shows = $followed_shows->where('state', '=', 3);
+                return Response::json(View::make('users.shows_cards', ['shows' => $completed_shows])->render());
+            } elseif ($inputs['state'] == 4) {
+                $followed_shows = $this->showRepository->getShowFollowedByUser($user->id);
+                $abandoned_shows = $followed_shows->where('state', '=', 4);
+                return Response::json(View::make('users.shows_abandoned_cards', ['shows' => $abandoned_shows])->render());
+            } elseif ($inputs['state'] == 5) {
+                $followed_shows = $this->showRepository->getShowFollowedByUser($user->id);
+                $to_see_shows = $followed_shows->where('state', '=', 5);
+                return Response::json(View::make('users.shows_cards', ['shows' => $to_see_shows])->render());
             }
         }
         return 404;
+    }
+
+    public function unfollowShow($show) {
+        $user = Auth::user();
+
+        if ($user->shows->contains($show)) {
+            $user->shows()->detach($show);
+        }
     }
 }
